requirement.txt
certifi==2022.6.15
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.3
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
urllib3==1.26.9
